[Fix] 앱 백그라운드에서의 버그 수정
앱 백그라운드에서 불러올 때 StampView가 다시 안그려지는 버그 수정
