+++ AsyncSwift/Managers/KeyChainManager.swift
@@ -5,7 +5,6 @@
 //  Created by Inho Choi on 2022/09/15.
 //
 
-import Foundation
 import UIKit
 
 final class KeyChainManager {
